Cleans for flake8 errors.
